+++ app.js
@@ -7,7 +7,7 @@ var logger = require('morgan');
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
 var routerRouter = require('./routes/router');
-var routerRouter = require('./routes/upload');
+var routerRouter = require('./routes/blob-upload');
 
 var app = express();
 
@@ -24,7 +24,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
 app.use('/router', routerRouter);
-app.use('/upload', routerRouter);
+app.use('/blob-upload', routerRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
