+++ views/layout.pug
@@ -16,7 +16,7 @@ html
             li
               a(href="/") ホーム
             li
-              a(href="/upload") アップロード
+              a(href="/blob-upload") AzureBlobアップロード
             li
               a(href="/router") ルーターサンプル
     block content
